save progress on audio streaming

+++ mic.go
@@ -3,8 +3,10 @@ package main
 import (
 	"context"
 	"log"
+	//	"math"
 
 	"github.com/figadore/go-intercom/pb"
+	//	"github.com/jfreymuth/pulse"
 )
 
 func startStreaming(ctx context.Context, stream pb.Receiver_AudioClient) error {
@@ -25,7 +27,7 @@ func startStreaming(ctx context.Context, stream pb.Receiver_AudioClient) error {
 
 func sendAudio(stream pb.Receiver_AudioClient) error {
 	bytes := pb.AudioData{
-		Data: []byte{4, 2, 0},
+		Data: []float32{4, 2, 0},
 	}
 	err := stream.Send(&bytes)
 	if err != nil {
@@ -33,3 +35,22 @@ func sendAudio(stream pb.Receiver_AudioClient) error {
 	}
 	return nil
 }
+
+//var t, phase float32
+//
+//func synth(out []float32) (int, error) {
+//	for i := range out {
+//		if t > 4 {
+//			return i, pulse.EndOfData
+//		}
+//		x := float32(math.Sin(2 * math.Pi * float64(phase)))
+//		out[i] = x * 0.1
+//		f := [...]float32{440, 550, 660, 880}[int(2*t)&3]
+//		phase += f / 44100
+//		if phase >= 1 {
+//			phase--
+//		}
+//		t += 1. / 44100
+//	}
+//	return len(out), nil
+//}
